requirements.txt
pydub
soundfile
noisereduce
pyloudnorm
mutagen
tqdm
openai-whisper
scipy
numpy
